# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Gym CRM API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/imports/templates/{entity}": {
      "get": {
        "tags": [
          "imports"
        ],
        "summary": "Template",
        "operationId": "template_api_imports_templates__entity__get",
        "parameters": [
          {
            "name": "entity",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ImportEntity"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/imports": {
      "get": {
        "tags": [
          "imports"
        ],
        "summary": "List Imports",
        "operationId": "list_imports_api_imports_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ImportJobOut"
                  },
                  "type": "array",
                  "title": "Response List Imports Api Imports Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "imports"
        ],
        "summary": "Upload",
        "operationId": "upload_api_imports_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_api_imports_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/imports/{job_id}": {
      "get": {
        "tags": [
          "imports"
        ],
        "summary": "Get Import",
        "operationId": "get_import_api_imports__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportJobOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/imports/{job_id}/check": {
      "post": {
        "tags": [
          "imports"
        ],
        "summary": "Check",
        "description": "Saves the column mapping and validates every row (nothing is written).",
        "operationId": "check_api_imports__job_id__check_post",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Job Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/imports/{job_id}/run": {
      "post": {
        "tags": [
          "imports"
        ],
        "summary": "Run",
        "operationId": "run_api_imports__job_id__run_post",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportJobOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/imports/{job_id}/problems.csv": {
      "get": {
        "tags": [
          "imports"
        ],
        "summary": "Problems",
        "operationId": "problems_api_imports__job_id__problems_csv_get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_upload_api_imports_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          },
          "entity": {
            "$ref": "#/components/schemas/ImportEntity"
          }
        },
        "type": "object",
        "required": [
          "file",
          "entity"
        ],
        "title": "Body_upload_api_imports_post"
      },
...
      "CheckIn": {
        "properties": {
          "mapping": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Mapping"
          },
          "duplicate_mode": {
            "$ref": "#/components/schemas/DuplicateMode",
            "default": "skip"
          },
          "date_order": {
            "$ref": "#/components/schemas/DateOrder",
            "default": "dmy"
          }
        },
        "type": "object",
        "required": [
          "mapping"
        ],
        "title": "CheckIn"
      },
      "CheckOut": {
        "properties": {
          "job": {
            "$ref": "#/components/schemas/ImportJobOut"
          },
          "new": {
            "type": "integer",
            "title": "New"
          },
          "duplicates": {
            "type": "integer",
            "title": "Duplicates"
          },
          "invalid": {
            "type": "integer",
            "title": "Invalid"
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/RowProblem"
            },
            "type": "array",
            "title": "Errors"
          },
          "preview": {
            "items": {
              "$ref": "#/components/schemas/PreviewRow"
            },
            "type": "array",
            "title": "Preview"
          }
        },
        "type": "object",
        "required": [
          "job",
          "new",
          "duplicates",
          "invalid",
          "errors",
          "preview"
        ],
        "title": "CheckOut"
      },
...
      "DateOrder": {
        "type": "string",
        "enum": [
          "dmy",
          "mdy",
          "ymd"
        ],
        "title": "DateOrder"
      },
...
      "DuplicateMode": {
        "type": "string",
        "enum": [
          "skip",
          "update"
        ],
        "title": "DuplicateMode"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ImportEntity": {
        "type": "string",
        "enum": [
          "members",
          "leads",
          "payments",
          "checkups",
          "plans"
        ],
        "title": "ImportEntity"
      },
      "ImportJobOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "entity": {
            "$ref": "#/components/schemas/ImportEntity"
          },
          "status": {
            "$ref": "#/components/schemas/ImportStatus"
          },
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "columns": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Columns"
          },
          "total_rows": {
            "type": "integer",
            "title": "Total Rows"
          },
          "mapping": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Mapping"
          },
          "duplicate_mode": {
            "$ref": "#/components/schemas/DuplicateMode"
          },
          "date_order": {
            "$ref": "#/components/schemas/DateOrder"
          },
          "created": {
            "type": "integer",
            "title": "Created"
          },
          "updated": {
            "type": "integer",
            "title": "Updated"
          },
          "skipped": {
            "type": "integer",
            "title": "Skipped"
          },
          "failed": {
            "type": "integer",
            "title": "Failed"
          },
          "processed": {
            "type": "integer",
            "title": "Processed"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "has_error_file": {
            "type": "boolean",
            "title": "Has Error File"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "entity",
          "status",
          "filename",
          "columns",
          "total_rows",
          "mapping",
          "duplicate_mode",
          "date_order",
          "created",
          "updated",
          "skipped",
          "failed",
          "processed",
          "error",
          "has_error_file",
          "created_at",
          "finished_at"
        ],
        "title": "ImportJobOut"
      },
      "ImportStatus": {
        "type": "string",
        "enum": [
          "uploaded",
          "ready",
          "running",
          "done",
          "failed"
        ],
        "title": "ImportStatus"
      },
...
      "PreviewRow": {
        "properties": {
          "row": {
            "type": "integer",
            "title": "Row"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "values": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Values"
          }
        },
        "type": "object",
        "required": [
          "row",
          "status",
          "values"
        ],
        "title": "PreviewRow"
      },
...
      "RowProblem": {
        "properties": {
          "row": {
            "type": "integer",
            "title": "Row"
          },
          "messages": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Messages"
          }
        },
        "type": "object",
        "required": [
          "row",
          "messages"
        ],
        "title": "RowProblem"
      },
...
      "UploadOut": {
        "properties": {
          "job": {
            "$ref": "#/components/schemas/ImportJobOut"
          },
          "sample": {
            "items": {
              "additionalProperties": {
                "anyOf": [
                  {
                    "type": "string"
                  },
                  {
                    "type": "null"
                  }
                ]
              },
              "type": "object"
            },
            "type": "array",
            "title": "Sample"
          },
          "suggested_mapping": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Suggested Mapping"
          }
        },
        "type": "object",
        "required": [
          "job",
          "sample",
          "suggested_mapping"
        ],
        "title": "UploadOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}